requirements.txt
Flask
Werkzeug
python-dotenv

matplotlib
opencv-python
pyzbar[scripts]

pdf2image
